Worked on Colours
Fixed the camera as it was bugged so some of the territories would be the wrong colour, it was originally fixed by kevin but he didn't commit it.

Added the hover and click colour changes.

Dont know what is up with it adding two files, since I didn't

In conjunction with issues 63,64,66.